implemented my front-end changes
This includes:
1. Creating a new folder called data_entry_page in the app directory. This folder now holds the file page.js, which is the dropdown menu with API implementation
2. The original page.js in the app directory now contains the code for the homepage with a link to the data entry page contained within the button
3. A new Public folder was created within the root directory. This folder will contain any files we might need to use while building the app. For now, it just has an images folder that has the current page background images.
