+++ package.json
@@ -14,9 +14,9 @@
     "@mui/material": "^6.1.2",
     "@mui/x-data-grid": "^7.18.0",
     "axios": "^1.7.7",
-    "next": "^14.2.13",
-    "react": "^18",
-    "react-dom": "^18"
+    "next": "^14.2.14",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
     "eslint": "^8",
